refactor: Refactor multi-session switch account UI
Split and reorder multi-session menu components: SwitchAccountItem now renders a submenu trigger (icon + label) that opens SwitchAccountMenu, while the new SwitchAccountMenuItem renders the per-session dropdown item (UserView + spinner) and performs setActiveSession. Updated multi-session plugin to use the submenu trigger, adjusted imports/exports and small formatting/type tweaks across related files.
