+++ packages/core/src/plugins/multi-session/multi-session-plugin.ts
