+++ packages/heroui/src/components/multi-session/switch-account-item.tsx
@@ -1,47 +1,44 @@
-import {
-  type MultiSessionAuthClient,
-  useAuth,
-  useSetActiveSession
-} from "@better-auth-ui/react"
-import { Dropdown, Spinner } from "@heroui/react"
-import type { Session, User } from "better-auth"
-import { UserView } from "../user/user-view"
+import { useAuthPlugin } from "@better-auth-ui/react"
+import { Persons } from "@gravity-ui/icons"
+import { Dropdown, Label } from "@heroui/react"
 
-type DeviceSession = {
-  session: Session
-  user: User
-}
+import { multiSessionPlugin } from "../../lib/multi-session/multi-session-plugin"
+
+import { SwitchAccountMenu } from "./switch-account-menu"
 
 export type SwitchAccountItemProps = {
-  deviceSession: DeviceSession
+  className?: string
 }
 
 /**
- * Render a dropdown item for switching to a different authenticated session.
+ * Render a dropdown menu item for switching between multiple authenticated sessions.
  *
- * @param deviceSession - The device session to display and switch to when pressed
- * @returns The switch account dropdown item as a JSX element
+ * This component renders as a submenu trigger that opens the switch account menu.
+ * It should be rendered inside a Dropdown.Menu.
+ *
+ * @param className - Optional additional CSS class names
+ * @returns The switch account menu item as a JSX element
  */
-export function SwitchAccountItem({ deviceSession }: SwitchAccountItemProps) {
-  const { authClient } = useAuth()
-  const { mutate: setActiveSession, isPending } = useSetActiveSession(
-    authClient as MultiSessionAuthClient,
-    {
-      onSuccess: () => window.scrollTo({ top: 0 })
-    }
-  )
+export function SwitchAccountItem({ className }: SwitchAccountItemProps) {
+  const { localization: multiSessionLocalization } =
+    useAuthPlugin(multiSessionPlugin)
 
   return (
-    <Dropdown.Item
-      className="px-2"
-      isDisabled={isPending}
-      onPress={() =>
-        setActiveSession({ sessionToken: deviceSession.session.token })
-      }
-    >
-      <UserView user={deviceSession.user} />
+    <Dropdown.SubmenuTrigger>
+      <Dropdown.Item
+        className={className}
+        textValue={multiSessionLocalization.switchAccount}
+      >
+        <Persons className="text-muted" />
 
-      {isPending && <Spinner color="current" size="sm" className="ml-auto" />}
-    </Dropdown.Item>
+        <Label>{multiSessionLocalization.switchAccount}</Label>
+
+        <Dropdown.SubmenuIndicator />
+      </Dropdown.Item>
+
+      <Dropdown.Popover className="min-w-40 md:min-w-56 max-w-[48svw]">
+        <SwitchAccountMenu />
+      </Dropdown.Popover>
+    </Dropdown.SubmenuTrigger>
   )
 }
