+++ packages/heroui/src/lib/multi-session/multi-session-plugin.ts
@@ -4,13 +4,13 @@ import {
 } from "@better-auth-ui/core/plugins"
 
 import { ManageAccounts } from "../../components/multi-session/manage-accounts"
-import { SwitchAccountMenuItem } from "../../components/multi-session/switch-account-menu-item"
+import { SwitchAccountItem } from "../../components/multi-session/switch-account-item"
 import type { AuthPlugin } from "../auth-plugin"
 
 export function multiSessionPlugin(options: MultiSessionPluginOptions = {}) {
   return {
     ...coreMultiSessionPlugin(options),
     accountCards: [ManageAccounts],
-    userMenuItems: [SwitchAccountMenuItem]
+    userMenuItems: [SwitchAccountItem]
   } satisfies AuthPlugin
 }
